ml-services/ML-Service/paymigo_assistant/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
httpx
openai
sentence-transformers
faiss-cpu
numpy
pandas
transformers
torch
pytest
